move i18n.js to config/lib
