+++ .eslintrc.js
@@ -115,6 +115,7 @@ module.exports = {
     // overrides for client/react code
     files: [
       'config/webpack/**',
+      'config/lib/i18n.js',
       'modules/core/client/app/config.js',
       'modules/**/client/components/**',
       'modules/core/client/directives/tr-boards.client.directive.js',
